Renamed THe getLine.h to getThatLine  due to some conflicts and some other things in the stuff.h
