+++ output.html
@@ -1,6 +1,27 @@
 <!DOCTYPE html>
 <html lang="en" dir="ltr">
   <head>
+    <style>
+      h3 {
+        color: orange;
+      }
+      #para1 {
+        text-align: left;
+        color: red;
+        margin-top: 100px;
+        margin-bottom: 100px;
+        marg1. this is a question</body></html>
+: 80px;
+      }
+      #para2 {
+        text-align: right;
+        color: black;
+        margin-top: 100px;
+        margin-bottom: 100px;
+        margin-right: 150px;
+        margin-left: 80px;
+      }
+</style>
     <meta charset="utf-8">
     <title>Question Paper</title>
   </head>
@@ -9,5 +30,16 @@
     <h3>Sample School Name</h3>
     <h4>Qustion Paper</h4>
     </div>
-    <div align  = "left">
-1. this is a question</body></html>
+    <div align  = "left" id="para1">
+      <p>Name:______________________________<br>
+        Class:______<br>
+      </p>
+    </div>
+    <div id="para2">
+      <p>
+        Subject:_____________________________<br>
+        Roll No:_____________________________<br>
+      </p>
+    </div>
+  </body>
+</html>
